Fix golang tests
