--- traffic_ops/traffic_ops_golang/ats/atsprofile/profile_test.go
@@ -1,52 +1,0 @@
-package atsprofile
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import "testing"
-
-func TestTrimParamUnderscoreNumSuffix(t *testing.T) {
-	inputExpected := map[string]string{
-		``:                         ``,
-		`a`:                        `a`,
-		`_`:                        `_`,
-		`foo__`:                    `foo__`,
-		`foo__1`:                   `foo`,
-		`foo__1234567890`:          `foo`,
-		`foo_1234567890`:           `foo_1234567890`,
-		`foo__1234__1234567890`:    `foo__1234`,
-		`foo__1234__1234567890_`:   `foo__1234__1234567890_`,
-		`foo__1234__1234567890a`:   `foo__1234__1234567890a`,
-		`foo__1234__1234567890__`:  `foo__1234__1234567890__`,
-		`foo__1234__1234567890__a`: `foo__1234__1234567890__a`,
-		`__`:                       `__`,
-		`__9`:                      ``,
-		`_9`:                       `_9`,
-		`__35971234789124`:         ``,
-		`a__35971234789124`:        `a`,
-		`1234`:                     `1234`,
-		`foo__asdf_1234`:           `foo__asdf_1234`,
-	}
-
-	for input, expected := range inputExpected {
-		if actual := trimParamUnderscoreNumSuffix(input); expected != actual {
-			t.Errorf("Expected '%v' Actual '%v'", expected, actual)
-		}
-	}
-}
